Shelve many "helm-*" packages
The author initiated their deprecation in
https://github.com/melpa/melpa/pull/9520, where he asked for their
removal from Melpa.  I waited a few months to see how that plays
out.  The removals resulted in many pull-request all over the place
to remove these packages from starter-kits and such, but it appears
nobody requested that any of these packages be added back to Melpa,
so presumably these are indeed obsolete.